Fix distribution editor inputs, header cramping, and settings visual issues
- Fix CSS specificity bug: .modal input[type=number] (specificity 0,2,1) was
  overriding .dist-edit-row input (0,1,1), making the distribution % inputs
  render at width:100% instead of the intended 70px fixed width, breaking
  the flex row layout. Added .modal .dist-edit-row input[type=number] (0,3,1)
  to definitively restore the correct sizing.

- Fix header cramping: add flex:1/min-width:0 to .brand so it fills available
  space; add flex-shrink:0 to .actions so buttons never squish the title;
  tighten button gap from 8px to 6px.

- Improve settings section contrast: reorder-row background changed from
  var(--card) to var(--bg-elev) so rows are visually distinct from the
  card container they sit inside.

- Shorten stats strip dates from month:'short' to month:'numeric' (e.g.
  "١٥/٦/٢٣" instead of "١٥ يونيو ٢٣") so they fit in the narrow 1/3-width
  cells without overflow. Reduce .sett-stat-v font-size to 12px.

- Bump SW cache to v11.

+++ app.js
@@ -2841,14 +2841,14 @@ function updateSettingsStats(){
   document.getElementById('statTxCount').textContent = state.transactions.length.toLocaleString('ar-EG');
   if(state.transactions.length){
     const first = state.transactions.reduce((min,t)=> t.ts<min.ts ? t : min, state.transactions[0]);
-    document.getElementById('statFirstTx').textContent = new Date(first.ts).toLocaleDateString('ar-EG', {day:'numeric', month:'short', year:'2-digit'});
+    document.getElementById('statFirstTx').textContent = new Date(first.ts).toLocaleDateString('ar-EG', {day:'numeric', month:'numeric', year:'2-digit'});
   } else {
     document.getElementById('statFirstTx').textContent = '—';
   }
   try{
     const last = localStorage.getItem(LS_PREFIX + 'lastEdit');
     document.getElementById('statLastEdit').textContent = last
-      ? new Date(parseInt(last)).toLocaleString('ar-EG', {day:'numeric', month:'short', hour:'2-digit', minute:'2-digit', hour12:false})
+      ? new Date(parseInt(last)).toLocaleString('ar-EG', {day:'numeric', month:'numeric', hour:'2-digit', minute:'2-digit', hour12:false})
       : '—';
   }catch(e){
     document.getElementById('statLastEdit').textContent = '—';
